Added Temaining Routes

+++ server.js
@@ -67,6 +67,12 @@ app.get("/api/candidate/:id", (req, res) => {
 
 // Delete a candidate
 app.delete("/api/candidate/:id", (req, res) => {
+  const errors = inputCheck(req.body, "party_id");
+
+  if (errors) {
+    res.status(400).json({ error: errors });
+    return;
+  }
   const sql = `DELETE FROM candidates WHERE id = ?`;
   const params = [req.params.id];
   db.run(sql, params, function (err, result) {
@@ -109,6 +115,74 @@ app.post("/api/candidate", ({ body }, res) => {
       message: "success",
       data: body,
       id: this.lastID,
+    });
+  });
+});
+
+// All Parties
+app.get("/api/parties", (req, res) => {
+  const sql = `SELECT * FROM parties`;
+  const params = [];
+  db.all(sql, params, (err, rows) => {
+    if (err) {
+      res.status(500).json({ error: err.message });
+      return;
+    }
+
+    res.json({
+      message: "success",
+      data: rows,
+    });
+  });
+});
+
+// Individual Party
+app.get("/api/party/:id", (req, res) => {
+  const sql = `SELECT * FROM parties WHERE id = ?`;
+  const params = [req.params.id];
+  db.get(sql, params, (err, row) => {
+    if (err) {
+      res.status(400).json({ error: err.message });
+      return;
+    }
+
+    res.json({
+      message: "success",
+      data: row,
+    });
+  });
+});
+
+// Delete a Party
+app.delete("/api/party/:id", (req, res) => {
+  const sql = `DELETE FROM parties WHERE id = ?`;
+  const params = [req.params.id];
+  db.run(sql, params, function (err, result) {
+    if (err) {
+      res.status(400).json({ error: res.message });
+      return;
+    }
+
+    res.json({ message: "successfully deleted", changes: this.changes });
+  });
+});
+
+// Update Candidates Party Affiliation
+app.put("/api/candidate/:id", (req, res) => {
+  const sql = `UPDATE candidates SET party_id = ? 
+                 WHERE id = ?`;
+  const params = [req.body.party_id, req.params.id];
+
+  db.run(sql, params, function (err, result) {
+    if (err) {
+      res.status(400).json({ error: err.message });
+      return;
+    }
+
+    res.json({
+      message: "success",
+      data: req.body,
+      changes: this.changes,
     });
   });
 });
